Add transaction_log_dir option
It allows to put the "LATEST.LOG" and "PREVIOUS.LOG" in a different directory
than the one used for all other Mnesia data. This way, the transaction logs can
be stored on a different filesystem for example, which may be beter suited for
them. This in turn can lead to improved performance.