Forbid nonsensical module names
When attempting to compile a module named `''` (in the file `.erl`),
the compiler would produce the following error message:

    1> c("").
    ..beam: Module name '' does not match file name '.'
    error

Compiling with `erlc` would also fail.

Instead of trying to fix this problem, this commit forbids nonsensical
module names. Here are the types of module names that are no longer
allowed:

* An empty module name (the atom `''`).

* Names containing control characters (16#00 through 16x1F and 0x7F
  through 16#9F).

* Names containing **only** spaces and soft hyphens.

We don't change the rules for function names (yet, anyway), because
they don't have the same fundamental problem that module names have,
namely that a file in the file system must be have the same base name
as the module.

Solves #6026