compiler: Add type info for bs_init_writable
The SSA instruction bs_init_writable is not known to the SSA type
analysis pass, nor is its result type known to the BEAM
validator. This patch adds type information for bs_init_writable and
as it allows the SSA type pass to elide is_bitstring tests, it also
extends the BEAM validator to consider the result of bs_init_writable
to be a bitstring.