kernel: Change -noshell and -noinput to use user_drv
We want as much of the I/O work as possible to go through
the user_drv as it will use dirty I/O schedulers to schedule
the work. Also this makes the unicode detection work even
for -noshell/-noinput type systems.

This commit also adds user_drv:start_shell/0 to allow the user
to start the shell after the fact. For instance when rebar3
starts as en escript it may want to start a shell depending on
what arguments are given to it.