Combine literal values in integer expressions
Consider:

    lowercase([H|T]) when is_integer(H), $A =< H, H =< $Z ->
        [H - $A + $a|lowercase(T)];
    lowercase([H|T]) ->
        [H|lowercase(T)];
    lowercase([]) ->
        [].

In the expression `H - $A + $a`, the compiler does not automatically
combine the two character literals to a single literal. For that to
happen, one would have to write the expression as `H - ($A - $a)`.

Implement a simple reassociation optimization that will combine
integer literals in integer expressions. (Note that the optimization
could be unsafe for expression involvolving floating point numbers.)