[erts] Fix lost exit signal due to connection loss
Loss of a connection over which a not yet completed spawn request was
ongoing could cause loss of exit signals. Such loss of exit signals was
very rare. Besides the above described connection loss also the following
conditions had to be satisfied:
* The spawn request that was interrupted by the connection loss also had
  to set up a link between the parent process and the child process.
* The parent process that issued the spawn request also had to be
  terminating while the spawn request was interrupted by the connection
  loss.
* The same parent process also had to have made other spawn requests to
  other nodes than to the node to which the connection was lost.
* These spawn requests to the other nodes also had to set up links.
* These spawn requests to the other nodes also had to be not yet completed
  at the time of the connection loss. That is, the spawn reply from the
  child process had not yet reached the parent process.
If all the conditions above were met, exit signals to the children spawned
due to the above described spawn requests to other nodes *could* be lost.

The above bug also caused a significant memory leak when it was triggered
since the destruction of the parent process never completed.