Dialyzer: Fix checking of behaviours
Fix two bugs in Dialyzer's behaviour checking:

* When a mandatory callback function is present but not exported,
Dialyzer would not complain about a missing callback. However, it
would complain if the arguments and/or return values were
incompatible.

* When an optional callback function was not exported and had
incompatible arguments and/or return values were incompatible,
Dialyzer would complain. This has been changed to suppress the
warning, because the function might not be intended to be a callback
function, for instance if a release added a new optional callback
function (such as `format_status/1` for the `gen_server` behaviour
added in OTP 25).