[crypto] Changed the behaviour of the engine load functions
The engine load/unload functions have got changed semantics to get
a more consistent behaviour and work correct when variables are
garbage collected.

The load functions now don't register the methods for the engine
to replace. That will now be handled with the new functions
engine_register/engine_unregister if needed.

Some functions functions are removed from the documentation and
therefor the API, but they are left in the code for compatibility.

engine_load/4: is now the same as engine_load/3
engine_unload/2: is now the same as engine_unload/1
ensure_engine_loaded/3: is now the same as ensure_engine_loaded/2
ensure_engine_unloaded/1, ensure_engine_unloaded/2: is now
   the same as engine_unload/1