ssl: Refactor handle_options
Split clauses to be seperate functions.

Rename expand_options/2 to split_options/2 which is more what it does.
Remove some already removed (deprecated) options which is handled earlier.

Explicit order and initilize options that other options are dependent on.
Collect functionality (mostly) in one place to make it easier to overview.

This commit have changed error messages for a couple of error cases,
and is also more restrict with checking the options.

{options, incompatible, a, b} => {options, incompatible, [a,b]}
{options, dependency, a,b} => {options, incompatible, [a,^b]}
{conflict_options, a, b} => {options, incompatible, [a,b]}