Enhance creation of maps with literal keys
An optimization in the loader was added in: https://github.com/erlang/otp/pull/1498

When a small map is created where all keys are literals, the key tuple
is created as a literal. That means that the key tuple can be shared
for all instances of the map created by the same code.

However, that does not always work, for example for code such
as:

    decimal(Int) ->
        #{type => decimal, int => Int, exp => 0}.

The compiler would try to be smart and rewrite the code to:

    decimal(Int) ->
        (#{type => decimal, exp => 0})#{int => Int}.

Expressed in that way, the loader optimization cannot be applied and
a new key tuple will be created every time `decimal/1` is called.

This commit removes that problematic transformation. The implementation
is based on José Valim's suggested implementation.

Fixes #6139