[kernel|pg] Add tagged processes feature
Added possibility to associate a process with a tag when joining pg
groups. This tag can be useful to identify particular processes within a
given group.

For example, suppose some resource is controlled by three processes -
each one controls one aspect of this resource. Let's name these aspects
top, middle and bottom.
These processes are added to a pg group. In this way, it's easy to get
the processes associated to a resource, but it's hard to identify the
process related to a specific aspect among the returned ones.
With tagged processes, these processes could be associated with some tag
in order to be easily identified within that group.

This use case may be executed as follow:
  %%%%
  %% The resource instance 42 is controlled by three processes (local or
  %% remote).
  %% Each process manages one aspect (top, middle and bottom) of that
  %% resource.
  %%%%

  %% Set up group
  Scope = some_scope,
  GroupId = 42,
  Top = new_pid(top, GroupId), % returns pid()
  Middle = new_pid(middle, GroupId), % returns pid()
  Bottom = new_pid(bottom, GroupId), % returns pid()
  Processes = [{top, Top}, {middle, Middle}, {bottom, Bottom}].
  pg:join(Scope, GroupId, Processes).

  %% Getting members
  Members = pg:get_members(Scope, GroupId).
  % something like [{top, Pid1}, {middle, Pid2}, {bottom, Pid3}]

  %% it's easy to get the process which controls the aspect "middle" in
  %% this group
  {middle, MiddlePid} = lists:keyfind(middle, 1, Members).