Eliminate internal consistency failure
When attempting to compile the following code:

    foo(V) when (V / V < (V orelse true)); V; V ->
        (V = (a =/= V)) orelse 0.

the `beam_validator` pass produces the following message:

    t:1: function foo/1+12:
      Internal consistency check failed - please report this bug.
      Instruction: {test,is_eq_exact,{f,9},[{x,1},{atom,true}]}
      Error:       {type_conflict,{t_atom,[true]},{t_atom,[a]}}:

The type conflict mentioned refers to the expression `a =/= V`. At
the point of its evaluation, `V` is known to `beam_validator` to be
the atom `true`; thus, evaluation of `a =/= V` is redundant because the
result is always `false`.

The root cause of the problem is that `beam_validator` has a stronger
type analysis for `=/=` expressions than `beam_ssa_type`. By
strengthening the type analysis for `=/=` in `beam_ssa_type`, the
offending redundant tests are removed and `beam_validator` is again
silent.

It turns out that stronger type analysis causes a new problem for
code such as:

    foobar(L) when L =/= [], is_list(L), L =/= [], hd(L) ->
        ok.

See the added comment and code in beam_types for a description of
the problem and its solution.

Fixes #6183