dialyzer: Fix unsound behaviour subtype checks
An invariant inside Dialyzer is that it is always safe to "simplify" a
type, (e.g. 1..1000 -> pos_integer(), or a | b | c | .. | z -> atom()).
This is safe when checking for overlap between specs and inferred types,
since the overlap never gets smaller when one of the types is grown, so
we don't make a situation where there's an overlap into one where
there's not, so we don't generate false positives.

For the case of behaviours, we use subtyping when checking the specs
and inferred types with respect to the behaviour definition. In this
case, the spec in source code may correctly be a subtype of the behaviour,
but due to internal simplifications which "should" always be safe, we can
make what was a subtype of the behaviour into something which is not,
and hence generate a false positive error about valid code.

The fix here is to switch behaviour checking to use overlap checking as
is used elsewhere in Dialyzer: it's a well tested codepath and is
consistent with the simplification logic. The drawback is that if it
weren't for the simplification logic (which is required to keep
performance reasonable), subtyping would be the correct thing to enforce,
so those cases where a spec does have an overlap with the behaviour, but
is not a subtype, will no longer be reported, since Dialyzer can't
safely tell that situation apart from valid code which has been
simplified.

We also add special casing to match the original intent, which allows
callbacks to crash (i.e. have a none() return type).