kernel: Fix loading of prim_tty nif in embedded mode
When running in embedded mode the on_load function is called
by the init process after the user processes are started. So
we add a way for user to call the on_load function earlier
in the boot sequence so that the shell can be started when
it should.

We cannot move the on_load calls for all modules to an earlier
place in the kernel boot sequence as the code in on_load may
depend on kernel services being started.