kernel: Incorporate "oldshell" into user_drv
Instead of oldshell using a completely separate implementation
we adjust user_drv to setup group to act as an oldshell when
needed.

If needed shell:start_interactive will now start the oldshell.
This is needed for rebar3 to be able to start as it should in
environments where TERM=dumb or we do not have TERMCAP.