ssl: Make sure crypto and asn1rt_nif are loaded at embedded start
When starting a node in embedded mode and using tls dist we
need to make sure that the proper nifs are loaded when the
dist is started. During a normal boot the on_load functions are
run when the kernel safe sup is started which is much later than
when the distribution is started.

Closes #6085