Change strategy for fetching callback mode
Instead of setting the callback mode to undefined at state machine start,
code change and callback module change, and then calling Module:callback_mode()
at the first event; now call Module:callback_mode() directly at those occasions.

This fixes a bug that the state_enter flag was updated too late
when changing callback modules so the flag value from the old
module was used for the first state function call in the new module.

Hence if the two modules used different state_enter flags,
the new module could be called with a state enter call which it was
unprepared for, or the new module could not get a state enter call
that it was depending on.

As a consequence the sys:system_event() {code_change,_,_} cannot
be generated since a code change is now impossible to detect,
and the {enter,State} event has been changed to{enter,Module,State}.
Also, a {module,Module,State} event has been added to show module changes.