dialyzer: Fix literal value being used instead of type
A missing translation of a literal atom from a record name to the
corresponding atom type meant Dialyzer/TypEr would crash due to
trying to pattern match on a `c_literal` where a type was
expected.

This doesn't happen often, because it requires specs to be set to be
ignored (which results in the table of records being empty, activating
this codepath where the first literal atom in a tuple corresponding to
the record name is not converted into a type and is erroneously left as
a value-level literal).