Allow arbitrary binary Logger formatters
Data from Logger formatters that is already in binary format is passed
through as-is to the handler, allowing for binary logging. Lists are
still assumed to possibly contain unicode data and will be converted
with unicode:characters_to_binary/1 as before.

This opens up for re-using the Logger framework as is with a simple
formatter that e.g. calls term_to_binary/1 for binary logging.