Allow local redefinition of built-in types
If a module defines a type and a new built-in type is added with the
same name, Dialyzer will allow the redefinition of that particular
release for one release only. That makes it difficult to write code
that is supposed to work for multiple releases of OTP.

For example, the OTP 24 release introduced the `nonempty_binary`
type. Modules that had their own definition would continue to work in
OTP 24 but not in OTP 25.

Always allow the name of a built-in type to be reused locally, in the
same way that a definition of a function with the same name as a BIF
will take precedence over the BIF. Starting from OTP 25, the built-in
types belong to the erlang module, so it will always be possible to
get the built-in definition if needed by prefixing the name with
`erlang:`.

Closes #6132