Optimize matching of literal integer segments
Consider this function:

    e(<<1:2,I:14>>) ->
        I.

The first two bits of the binary must be equal to one for the match to
succeed. When a segment can have only one possible value, the compiler
will emit a `bs_match_string` instruction:

    {test,bs_start_match3,{f,1},1,[{x,0}],{x,1}}.
    {bs_get_position,{x,1},{x,0},2}.
    {test,bs_match_string,{f,3},[{x,1},2,{string,<<"@">>}]}.
    {bs_match,{f,3},
              {x,1},
              {commands,[{ensure_exactly,14},
                         {integer,2,{literal,[]},14,1,{x,2}}]}}.

The JIT will translate the `bs_match_string` instruction to a call
to a helper function.

To allow the JIT to generate more efficient code, change the compiler
to incorporate the matching of the first two bits into the `bs_match`
instruction:

    {test,bs_start_match3,{f,1},1,[{x,0}],{x,1}}.
    {bs_get_position,{x,1},{x,0},2}.
    {bs_match,{f,3},
              {x,1},
              {commands,[{ensure_exactly,16},
                         {'=:=',nil,2,1},
                         {integer,2,{literal,[]},14,1,{x,2}}]}}.
    {move,{x,2},{x,0}}.

In the `{'=:=',nil,2,1}` sub instruction, the `2` is the number of
bits to extract, and the `1` is the value to match. (The `nil` is
there for consistency -- other sub instructions have the number of
live registers in that position.)

The JIT will emit inlined code for this `bs_match` instruction, with
no calls to helper functions.