Improve type inference from binary matching
Consider the following example:

    m(Bin) ->
        <<I:8>> = Bin,
        I + 1.

The compiler's range analysis can figure out that `I` must be a
non-negative integer less than 256. With that information, the JIT
generates better code for the `+` operation (no type tests or overflow
tests are needed).

Extend the range analysis to also figure out the range for `I` when
the size is variable but with a known range. For example:

    m(Bin, Size) when Size =< 8 ->
        <<I:Size,_/bits>> = Bin,
        I + 1.

With the enhanced range analysis introduced in this commit, the
generated code will look like this:

    {test,is_ge,{f,1},[{integer,8},{x,1}]}.
    {test,bs_start_match3,{f,3},2,[{x,0}],{x,2}}.
    {test,bs_get_integer2,
          {f,3},
          3,
          [{tr,{x,2},{t_bs_context,1}},
           {tr,{x,1},{t_number,{'-inf',8}}},
           1,
           {field_flags,[{anno,[5,{file,"t.erl"}]},unsigned,big]}],
          {x,1}}.
    {gc_bif,'+',{f,0},2,[{tr,{x,1},{t_integer,{0,255}}},{integer,1}],{x,0}}.