jit: Fix erts_beamasm symbols
The symbols for erts_beamasm did not take into consideration
that there was not prologue for functions in that jit internal
module. This causes perf to show $erts_beamasm:normal_exit/0-CodeInfoPrologue
instead of $erts_beamasm:normal_exit/0 for the top frame of all
processes which is incorrect and possibly confusing.