Fix when ASN.1 module name does not match file
When a module identifier does not match the file name, the `.hrl` and the `.erl` files are generated with the same name as the input file. However, the include directive in the `.erl` file uses the module identifier, resulting in a compilation error. This fixes the issue.

See example error, the module identifier in the ASN file is `Modkeystoretypes.hrl`:

```
erlc mod_keystore_types.asn
/src/mod_keystore_types.erl:8:10: can't find include file "Modkeystoretypes.hrl"
%    8| -include("Modkeystoretypes.hrl").
%     |          
```