erts: Decouple use of single-mapped memory from `perf` support
Running the JIT with single-mapped RWX memory for JIT:ed machine code
is needed for `perf` support. But as it is also useful for running
under QEMU user mode emulation [1] this patch adds a new flag
`+JMsingle bool` which controls the use of single-mapped RWX memory
without triggering the output of perf-related metadata.

The naming of the flag is intended to follow the (perceived) pattern
introduced by the `+JPperf`-flag, that is, `J` is a JIT-related option
and `M` stands for memory related things.

Thanks to Björn Gustavsson <bjorn@erlang.org> for suggesting how to
deal with Apple-silicon aarch64 and how to make the smoke test work on
Windows.

Personally I use the additional patch below in order to use the same
file system image for development on an amd64 build host and for
deployment to an aarch64-target board. The patch detects when the VM
is running using QEMU and if so, enables single-mapped memory. As the
patch only works on Linux and doesn't try to identify the version of
QEMU (if the bug gets fixed), I don't think it belongs on OTP
master. But as it is nevertheless useful, I include it here:

```
diff --git a/erts/emulator/beam/erl_init.c b/erts/emulator/beam/erl_init.c
index 42216da0b2..6f11c7586b 100644
--- a/erts/emulator/beam/erl_init.c
+++ b/erts/emulator/beam/erl_init.c
@@ -2408,6 +2408,40 @@ erl_start(int argc, char **argv)
 	erts_usage();
     }

+
+#if defined(HAVE_LINUX_PERF_SUPPORT)
+    /* We want to detect if we are running under QEMU user mode
+       emulation in order to activate single-mapped RWX memory for the
+       JIT:ed code. This only works on Linux, so we use
+       HAVE_LINUX_PERF_SUPPORT to conditionally compile this piece of
+       detection code.
+
+       To detect QEMU user mode we make use of the fact that our
+       parent process will always be QEMU, we detect that by looking
+       up our parent process and then checking if the final path
+       component of the symlink /proc/<parent>/exe starts with
+       "qemu-".
+    */
+    {
+        char symlink_buf[21 + 11]; /* space for any 64 bit integer
+                                      + /proc/%d/exe and terminator */
+        char target_buf[MAXPATHLEN];
+        ssize_t l;
+        erts_snprintf(symlink_buf, sizeof(symlink_buf),
+                      "/proc/%d/exe", getppid());
+        l = readlink(symlink_buf, target_buf, sizeof(target_buf));
+        if (l > 0 && l != sizeof(target_buf)) {
+            char *last_path_separator;
+
+            target_buf[l] = 0;
+            last_path_separator = memrchr(target_buf, '/', l);
+            if (last_path_separator && strncmp(last_path_separator + 1,
+                                               "qemu-", 5) == 0)
+                erts_jit_single_map = 1;
+        }
+    }
+#endif
+
 /* Output format on windows for sprintf defaults to three exponents.
  * We use two-exponent to mimic normal sprintf behaviour.
  */
```

[1] There is a bug in QEMU,
    see https://gitlab.com/qemu-project/qemu/-/issues/1034