Optimize is_binary/1 and is_bitstring/1 tests
The tests for is_binary/1 and is_bitstring/1 currently calls a common
helper function, which results in some extra instructions and
precludes some optimizations.

If it is known that the argument for is_binary/1 is a bitstring,
the test to distinguish a sub binary from a heap or refc binary
can be reduced to testing a single bit.

When the is_bitstring/1 test no longer shares code with is_binary/1,
it can be simplified because we no longer cares about which kind
of binary it is (sub, heap, or refc). It is sufficent to compare the
two bits in the header that distinguishes binaries from all other
terms.