Optimize is_nil/1 test
For x86_64, the `cmp` instruction can be shortened by one byte by
only comparing the least significant byte of the header word with
the NIL constant.

For AArch64, if the type is known to be either a cons or NIL,
one instruction can be eliminated by doing a bit test.