Dialyzer: Eliminate crash for code generated by Elixir
e55d8a9b209140 refactored dialyzer_dataflow and removed a clause that
could never be reached when analysing code generated by the standard
Erlang compiler. Unfortunately, it can be reached if the code was
generated by a parse transform or by another code generator such as
the Elixir compiler.

Here is the Elixir code that caused the crash.

https://github.com/ash-project/ash/blob/fe596db0b86ea58701e00523a0be408487bb8f27/lib/ash/query/operator/basic.ex#L66-L68

Closes #6323