[kernel|esock|test] Make ttest group be part of the standard test
The ttest (esock benchmark) group of the socket tes suite did not run.
Because it was not part of the standard test run, it hade broken and
we did not know it.
But because it contained so many test cases it took to much time to
run in its entirety. So, each test now only runs for 1 sec (instead of
10 sec). Also only 1/3 of the test cases will be run in a standard
run, the 'small' (ttest-) test cases.
The 'medium' and 'large' (ttest-) cases has to explicitly configured
to run.