Improve reuse of registers
Keep track of two memory locations and their associated registers
instead of just one, and also recognize more neutral instructions that
will not force the cache to be discarded. For the x86_64, the
`cmp_arg()` helper has been enhanced to avoid reloading BEAM
registers.

For the x86_64 JIT, with this commit 87361 loads of BEAM registers
are avoided, compared to 63360 loads before this commit.

For the AArch64 JIT, with this commit 13859 loads of BEAM registers
are avoided, compared to 9516 loads before this commit.