Rewrite the listen_port_options/1 test case
The test most probably did not test what was intended.
Since {reuseaddr,true} is default for the distribution
listen socket the test relied on the Unix semantics of not
allowing a second listening socket on the same port while
there exists an active listening socket on that port.

That semantics differ on Windows where {reuseaddr,true}
apparently allows creating a duplicate active listening socket.

The new test instead creates a connection and closes the
listening socket.  Now the server port can be reused
if {reuseaddr,true} is applied, which is the default.

Instead of trying to use an occupied port and verify
that it fails (which did not work on Windows), create
a connection and check by inspecting the result of
net_kernel:node_info/1 that the options for setting
listen port managed to do their thing.