[kernel|test] Fixed the fdconnect test case
Creating a socket (via gen_tcp:connect) using an already
existing file descriptor in principle requires descriptor to be
bound (its assumed to be, but the inet-driver attempt to verify).
The inet-driver does some checks (calls sockname) to ensure that
this is the case (and that its actually a 'socket' descriptor).
The problem is that for unix this function works even if the
descriptor is *not* bound. *But* this is not the case for Windows
(requires it to be bound).
The test case only created the descriptor but did not bind it,
so the test case worked on unix but not on windows.
The has now been fixed by providing the ifaddr-option to the
connect function causing an explicit bind.