AArch64: Eliminate crash in DEBUG build
In the DEBUG build, the contents of the X registers are either in
memory or in CPU registers, never both. After entering run-time
mode they are in memory, and their corresponding backing CPU
registers contains THE_NON_VALUE.

Therefore, we must enter run-time after the last use of the `ctx.reg`
register (which could a CPU register corresponding to a X register).