[snmp|test] Try limit the inet_backend groups on slow machines
When the test wuite is run with test_inet_backends = true, the
number of test cases almost tripples. On slow machines, this
can become a problem. So, on slow machines ('factor' > "limit"),
be skip groups: never the inet_backend_socket, but inet_backend_inet
and inet_backend_default can be skipped (depending on the 'factor').