stdlib: Have erl_scan recognize SSA check markers
Emit a `%ssa%` token when a block of lines all starting with the
string `"%ssa%"` is encountered. `%ssa%` tokens are only emitted when
the scanner is given the `checks` option. To allow for automatic
formatting, both `"%%ssa%` and `"%%%ssa%` are recognized as `%ssa%`
tokens.

Recognizing blocks of SSA checks requires that the scanner keeps track
of the fact that is scanning a check block across multiple invocations
of `tokens/{3,4}`. This is done by adding an additional field,
`in_check`, to `#erl_scan{}`.

Previously the value of #erl_scan{} was never changed by the scanner
function (the 6:th argument to `tokens1/7`), but as we now change the
`in_check` field in response to starting a check block, we have to
include the updated `#erl_scan{}` when the scanner function returns
`{more, ...}`.